refactor(component): 화면단위의 컴포넌트 위치변경
