+++ app/src/main/java/nextstep/shoppingcart/MainActivity.kt
@@ -4,7 +4,7 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import nextstep.shoppingcart.data.repository.ProductRepositoryImpl
-import nextstep.shoppingcart.ui.component.ProductListScreen
+import nextstep.shoppingcart.ui.ProductListScreen
 import nextstep.shoppingcart.ui.theme.ShoppingCartTheme
 
 class MainActivity : ComponentActivity() {
